# Faster conversion to dense matrix