--- a/test/sparsevector.jl
+++ b/test/sparsevector.jl
@@ -59,6 +59,7 @@ end
         @test isa(Array(x), Vector{Float64})
         @test Array(x) == xf
         @test Vector(x) == xf
+        @test convert(x) == xf
     end
 end
 @testset "show" begin