Ref #312 @ViralBShah 
(I also added a test to cover the `collect` - calls)

```julia
using SparseArrays, BenchmarkTools
for n in [16^i for i in 1:3]
    print("n = $n: ")
    sm = sprand(n, n, 0.2)
    @btime collect($sm)
end
# main
n = 16:   367.136 ns (1 allocation: 2.12 KiB)
n = 256:   64.300 μs (2 allocations: 512.05 KiB)
n = 4096:   45.581 ms (2 allocations: 128.00 MiB)
# PR
n = 16:   261.872 ns (1 allocation: 2.12 KiB)
n = 256:   23.200 μs (2 allocations: 512.05 KiB)
n = 4096:   27.871 ms (2 allocations: 128.00 MiB)
```
Edit: Benchmark updated for changing `copyto!` instead

<details>
<summary>versioninfo</summary>

```julia
Julia Version 1.8.5
Commit 17cfb8e65e (2023-01-08 06:45 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 8 × 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-13.0.1 (ORCJIT, tigerlake)
  Threads: 1 on 8 virtual cores
Environment:
  JULIA_PKG_OFFLINE = false
  JULIA_PKG_USE_CLI_GIT = true
```
</details>