@KnutAM 2023-01-21
> LGTM. What about the timings with the latest changes?

Unfortunately, I couldn't see any significant difference (which surprised me). 
```julia
A = Matrix{Float64}(undef, 100, 100);
S = sprand(100,100,0.2);
@btime copyto_1!($A, $S); # latest change
  2.022 μs (0 allocations: 0 bytes)
@btime copyto_2!($A, $S); # before latest change
  2.056 μs (0 allocations: 0 bytes)
@btime copyto_3!($A, $S); # latest change + inbounds
  1.770 μs (0 allocations: 0 bytes)
```