@dkarrasch 2023-01-26
> Unfortunately, I couldn't see any significant difference (which surprised me).

Well, but it's not that many nonzeros that you need to copy, about 2k. I think we should do whatever is safe and whatever does not make it worse, but simpler. Last thing: I believe we can add the inbounds annotation, the ingredients for the two loops and the `setindex!` look safe.