@ViralBShah 2023-01-26
Yeah try the benchmarks on something like sprand(1000, 1000, 0.5).