@KnutAM 2023-01-27
Thank's for the tip!
I'm not sure why `@inbounds` doesn't help as much on this one, perhaps branch prediction?
```julia
n=1000; d=0.5;
 A = zeros(n,n); S=sprand(n,n,d);
julia> @btime copyto_1!($A, $S); # Current state
  795.100 μs (0 allocations: 0 bytes)

julia> @btime copyto_2!($A, $S); # Earlier (num_rows*(col-1) + row)
  792.900 μs (0 allocations: 0 bytes)

julia> @btime copyto_3!($A, $S); # Current state + inbounds
  775.700 μs (0 allocations: 0 bytes)
  ```
  
  <details>
  <summary> Function definitions </summary>
  
  ```julia
julia> function copyto_1!(A::Array{T}, S::SparseMatrixCSC{<:Number}) where {T<:Number}
           isempty(S) && return A
           length(A) < length(S) && throw(BoundsError())

           # Zero elements that are also in S, don't change rest of A
           @inbounds for i in 1:length(S)
               A[i] = zero(T)
           end
           # Copy the structural nonzeros from S to A using
           # the linear indices (to work when size(A)!=size(S))
           num_rows = size(S,1)
           rowval = getrowval(S)
           nzval = getnzval(S)
           linear_index_col0 = 0   # Linear index before column (linear index = linear_index_col0 + row)
           for col in 1:size(S, 2)
               for i in nzrange(S, col)
                   row = rowval[i]
                   val = nzval[i]
                   A[linear_index_col0+row] = val
               end
               linear_index_col0 += num_rows
           end
           return A
       end
copyto_1! (generic function with 1 method)

julia> function copyto_2!(A::Array{T}, S::SparseMatrixCSC{<:Number}) where {T<:Number}
           isempty(S) && return A
           length(A) < length(S) && throw(BoundsError())

           # Zero elements that are also in S, don't change rest of A
           @inbounds for i in 1:length(S)
               A[i] = zero(T)
           end
           # Copy the structural nonzeros from S to A using
           # the linear indices (to work when size(A)!=size(S))
           num_rows = size(S,1)
           rowval = getrowval(S)
           nzval = getnzval(S)
           for col in 1:size(S, 2)
               for i in nzrange(S, col)
                   row = rowval[i]
                   val = nzval[i]
                   A[num_rows*(col-1)+row] = val
               end
           end
           return A
       end
copyto_2! (generic function with 1 method)

julia> function copyto_3!(A::Array{T}, S::SparseMatrixCSC{<:Number}) where {T<:Number}
           isempty(S) && return A
           length(A) < length(S) && throw(BoundsError())

           # Zero elements that are also in S, don't change rest of A
           @inbounds for i in 1:length(S)
               A[i] = zero(T)
           end
           # Copy the structural nonzeros from S to A using
           # the linear indices (to work when size(A)!=size(S))
           num_rows = size(S,1)
           rowval = getrowval(S)
           nzval = getnzval(S)
           linear_index_col0 = 0   # Linear index before column (linear index = linear_index_col0 + row)
           for col in 1:size(S, 2)
               @inbounds for i in nzrange(S, col)
                   row = rowval[i]
                   val = nzval[i]
                   A[linear_index_col0+row] = val
               end
               linear_index_col0 += num_rows
           end
           return A
       end
copyto_3! (generic function with 1 method)
  ```
</details>

> Last thing: I believe we can add the inbounds annotation, the ingredients for the two loops and the setindex! look safe.

I can break the `@inbounds` assumption:
```julia
A = zeros(n,n); S = sprand(n,n,d);
S[end,end] = 1.5
S.rowval[end] += 1 # Doing something stupid... (alternatively change S.colptr to assign some potentially random value)
copyto_1!(A,S); # BoundsError: attempt to access 1000×1000 Matrix{Float64} at index [1000001]
copyto_3!(A,S); # no error, can't this corrupt memory? E.g.
function test(A) # Deliberately looking at A[end+1]
    @inbounds begin
        a = A[length(A)+1]
    end
    return a
end
julia> A[end]
0.0
julia> test(A)
1.5
```

Maybe I understand [it](https://docs.julialang.org/en/v1/base/base/#Base.@inbounds) wrong, but I thought that without checking for such issues, I shouldn't use  `@inbounds`?