@dkarrasch 2023-01-30
Ok, if it doesn't help much then leaving it is fine, I think. OTOH, if someone messes with the internal representation after construction, then I'm not sure we can guarantee anything. I wonder if that would also break other code.